backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
alembic>=1.12.0
python-dotenv>=1.0.0
bcrypt==4.1.2
passlib==1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6